third_party/WebKit/LayoutTests/platform/win/fast/history/scroll-restoration/scroll-restoration-scale-not-impacted-expected.txt

CONSOLE WARNING: Upcoming versions will block content-initiated top frame navigations to data: URLs. For more information, see https://goo.gl/BaZAea.
This is a testharness.js-based test.
PASS Setting scrollRestoration to manual should not affect scale restoration 
Harness: the test ran to completion.
